добавление всех методов согласно задания в сервис, добавление логики + 2 первых @GetMapping

+++ game/src/main/java/com/game/repository/PlayerRepository.java
@@ -3,6 +3,6 @@ package com.game.repository;
 import com.game.entity.Player;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
 
 }
